Add code coverage example
This commit adds code coverage (CodeCov) support and integrates it
with GitHub CI so that every commit to the main trunk will trigger a
test coverage to be run on the GitHub cloud.  The coverage report is
then uploaded to CodeCov website.

Additionally, two badges (.svg images) are shown in README.md to
provide a quick glance at the overall CIDI and test coverage progress.